Neovim: Move lazy.nvim setup into 'config.lazy' module
Part of the reason for moving the configuration files into the `config`
directory was so that the `require 'lazy'` call didn't conflict with
this new `lazy.lua` file. So by having my setup of lazy in the
`config.lazy` module, it avoids this naming conflict.

In an additional effort to clean up the lazy file, all git related
plugins have been moved to `git.lua` and reqired in `config.lazy` here.