nix: Enable linux-builder on Darwin
While some of the setup was done here, there was still the tedious
process of creating the ssh config and adding the builder to the nix
settings so that it knew about the linux-builder. While going through
the process of automating this through the nix config, it turns out that
there is nix-darwin module to enable the linux-builder VM and have it
run in the background as a launchd service. This is even better than
what I was going to do because I was still going to manually run the
linux-builder VM everytime I wanted to use it.