nix: Use latest version of jujutsu
Updates jujutsu to the latest version (0.29.0) instead of (0.23.0) which
was missing features such as signing.

Additionally, the jujutsu config from home manager installs the jj
config in a location that has been deprecated since 0.23.0, on macOS so
we'll now link to the jj config using our symlink out of the nix store.