nix: Add aseprite derivation
Builds aseprite from source for aarch64-darwin. It requires using a
custom build of the skia fork by aseprite.

Moveover, in order to build it for aarch64-darwin, the clang++ compiler
that comes with Apple SDK and Xcode had to be used in order for this
to be compiled. Some of the directories to the CXX and OSX_SYSROOT for
CMake were hardcoded just for ease here. I would like to improve this in
the future to version of clang from the nix store, but for the sake of
getting this working, it's fine for now.

The final output is modified slightly just so that `aseprite` can be ran
from the command line, so the `share/aseprite/data` directory has been
moved to the `bin` directory so that `aseprite` can be globally ran from
the command line.